Cleanup: Git -> RCS
RELEASE_NOTES=n/a
