+++ internal/backend/storage/kv/inmem/store.go
@@ -128,14 +128,6 @@ func (m *InMem) Path() string {
 	return "inmem"
 }
 
-// Available will check if this backend is useable
-func (m *InMem) Available(ctx context.Context) error {
-	if m.data == nil {
-		return fmt.Errorf("not initialized")
-	}
-	return nil
-}
-
 // Fsck always returns nil
 func (m *InMem) Fsck(ctx context.Context) error {
 	return nil
