+++ internal/backend/storage/kv/ondisk/loader.go
@@ -37,11 +37,13 @@ func (l loader) Open(ctx context.Context, path string) (backend.RCS, error) {
 }
 
 // Clone loads an existing ondisk repo
-// WARNING: DOES NOT SUPPORT CLONE (yet)
 func (l loader) Clone(ctx context.Context, repo, path string) (backend.RCS, error) {
 	be, err := New(path)
 	debug.Log("Using RCS Backend %p: %s", be, be.String())
 	if err := be.SetRemote(ctx, repo); err != nil {
+		return nil, err
+	}
+	if err := be.Pull(ctx, "", ""); err != nil {
 		return nil, err
 	}
 	return be, err
